engine: fix gcc errors
1. Fixes GCC warnings that do not deserve me writing an indepth
explanation on what needed to be changed.

2. The original mask calculation for OPCODE_MOV makes zero sense because
the bit width for operands[0] is always > 31. I don't know what I was
thinking when I wrote this. The mask is now set to cover 32 and 64-bit
immediate values.

+++ src/bal_decoder.c
@@ -7,7 +7,7 @@ const bal_decoder_instruction_metadata_t *
 bal_decode_arm64(const uint32_t instruction)
 {
     // Index is top 11 bits
-    const uint16_t index = instruction >> 21;
+    const uint16_t index = (uint16_t)(instruction >> 21);
 
     const decoder_bucket_t *bucket = &g_decoder_lookup_table[index];
     for (size_t i = 0; i < bucket->count; ++i)
